Extract ViewState as common entity
