+++ app/src/main/java/io/horizontalsystems/bankwallet/entities/ViewState.kt
@@ -1,0 +1,6 @@
+package io.horizontalsystems.bankwallet.entities
+
+sealed class ViewState {
+    object Error : ViewState()
+    object Success : ViewState()
+}
